+++ eval.cu
@@ -139,7 +139,7 @@ int compare_result(
 // Returns 0 for success, -1 for failure.
 int eval(
     const char *text, int text_length,
-    std::string &pattern, int16_t pattern_length,
+    std::string &pattern, int pattern_length,
     StringMatchingFunction search_function,
     StringMatchingFunction reference_function = nullptr,
     bool verbose = false, int max_output_cnt = 100)
@@ -156,7 +156,7 @@ int eval(
     }
 
     std::vector<int> output(max_output_cnt);
-    std::vector<int16_t> fail(pattern_length + 1);
+    std::vector<int> fail(pattern_length + 1);
 
     auto start = std::chrono::steady_clock::now();
     int cnt = search_function(
